Create, Read methods are implemented

+++ database/migrations/2021_10_13_154933_create_food_table.php
@@ -17,7 +17,7 @@ class CreateFoodTable extends Migration
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->unsignedInteger('price');
+            $table->decimal('price', 10, 2);
             $table->foreignId('restaurant_id');
             $table->timestamps();
         });
